feat: react-scriptsでビルドできるようにしたビルドできるようにした
