+++ src/index.js
@@ -1,48 +1,9 @@
-const createElement = (type, props, ...children) => {
-  return {
-    type,
-    props: {
-      ...props,
-      children: children.map((child) =>
-        typeof child === 'object' ? child : createTextElement(child)
-      ),
-    },
-  };
-}
+import Ownact from './ownact';
 
-const createTextElement = (text) => {
-  return {
-    type: "TEXT_ELEMENT",
-    props: {
-      nodeValue: text,
-      children: []
-    }
-  };
-}
+const container = document.getElementById('root');
 
-const render = (element, container) => {
-  const dom = element.type === 'TEXT_ELEMENT'
-    ? document.createTextNode('')
-    : document.createElement(element.type);
-
-  const isProperty = key => key !== 'children';
-  Object.keys(element.props)
-    .filter(isProperty)
-    .forEach((name) => {
-      dom[name] = element.props[name];
-    });
-
-  element.props.children.forEach((child) => render(child, dom));
-
-  container.appendChild(dom);
-};
-
-const Ownact = {
-  createElement,
-  render,
-};
-
-/** @jsxImportSource Ownact */
+/** @jsxRuntime classic */
+/** @jsx Ownact.createElement */
 const element = (
   <div id="foo">
     <a>bar</a>
@@ -50,5 +11,4 @@ const element = (
   </div>
 );
 
-const container = document.getElementById('root');
 Ownact.render(element, container);
